requirements.txt
openai>=1.2.0
Flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pdf2image==1.17.0